version 1.1.2 Nextcloud 28 only

+++ lib/Migration/Version000000Date_20230228_154000.php
@@ -30,7 +30,14 @@ class Version000000Date_20230228_154000 extends SimpleMigrationStep
 
         $table = $schema->getTable('yumisign_nxtc_sess');
 
-        $table->dropColumn('expiry_date_lower');
+        /**
+         * ###### Fri Mar 15 07:48:10 CET 2024
+         * Add check phase before altering table => issue reported for RCDevs Support
+         * Cause : maybe manual DB updates/deletions
+         */
+        if ($table->hasColumn('expiry_date_lower')) {
+            $table->dropColumn('expiry_date_lower');
+        }
 
         return $schema;
     }
